+++ app/zzMetaModules/mmServer404.php
@@ -33,7 +33,7 @@ class moduleMetaServer404 extends \xan\moduleMeta {
 	
 	
 	///////////////////////////////////////////////////////////
-	// Functions Required by \xan\moduleMeta
+	// Abstract Functions Required by \xan\moduleMeta
 	
 	public function getDisplayList( \xan\recs $recs ) {
 		$code = $this->getDisplayName( $recs );
@@ -45,17 +45,6 @@ class moduleMetaServer404 extends \xan\moduleMeta {
 		$code = '';
 		$code = trim( $code );
 		return $code;
-	}
-	
-	public function getListItem( $idPrefix, \xan\recs $recs, $onClick ) {
-		$code = '';
-		return $code;
-	}
-	
-	public function getColLabel( $colName ) {
-		// Get Col Ele Meta
-		$colEle = $this->getColMeta( $colName, ELE_AS_LABEL );
-		return $colEle->colLabel;
 	}
 	
 	public function getColMeta( $colName, $typeAs = ELE_AS_DEFINED ) {
@@ -74,10 +63,13 @@ class moduleMetaServer404 extends \xan\moduleMeta {
 		$colMeta->choicesClearLabel = STR_CLEAR; // Add Clear
 		$colMeta->choicesOtherLabel = STR_OTHER; // Add Other
 		
+		// Sizes
+		$colMeta->widthForTable = ''; // No Default for Overrides Closer to Renderers
+		
 		// Columns Specifics
 		switch ( $colName ) {
 			case 'NULL':
-				$colEle->colLabelEN = 'NULL';
+				$colMeta->colLabelEN = 'NULL';
 				break;
 		}
 		
@@ -86,13 +78,6 @@ class moduleMetaServer404 extends \xan\moduleMeta {
 		
 		// Return the Element
 		return $colMeta;
-	}
-	
-	public function getColEleRender( $colName, $typeAs, \xan\tags $tags, \xan\recs $recs, \xan\formTag $formTag, \xan\response &$resp ) {
-		// Get Col Ele Meta
-		$colMeta = $this->getColMeta( $colName, $typeAs );
-		$code = \xan\eleDBMetaRender( $colMeta, $tags, $recs, $formTag, $resp );
-		return $code;
 	}